Miklos fixed some linux quotation command line bugs.  From Miklos:
While I updated to the latest tunit source I came across of the Runtime.getRuntime().exec(cmd) problem. You see, if you pass a string to this command, then it is first tokenized with the default tokenizer and broken into command+arguments. The problem is that your argument contains quotes, so the spaces inside the quotes get also broken up, and the make command on linux barfs on this, especially it says: make no command -D -T -U -I from the "... fifoqueue
-DTUNIT_TOTAL_NODES=1   " part of the argument.